// api/ref-data/models/ref-data-hmcts-service.model.ts
export interface RefDataHMCTSService {
  jurisdiction: string;
  service_id: number;
  org_unit: string;
  business_area: string;
  sub_business_area: string;
  service_description: string;
  service_code: string;
  service_short_description: string;
  ccd_service_name: string;
  last_update: string;
  ccd_case_types: string[];
}